initial qdrant changes

+++ backend/open_webui/apps/retrieval/vector/dbs/qdrant.py
@@ -1,4 +1,6 @@
-from qdrant_client import QdrantClient as Client, models
+from typing import Optional
+
+from qdrant_client import QdrantClient as Client
 from qdrant_client.http.models import (
     Distance,
     PointStruct,
@@ -7,9 +9,7 @@ from qdrant_client.http.models import (
     FieldCondition,
     MatchValue,
 )
-
-
-from typing import Optional
+from qdrant_client.models import models
 
 from open_webui.apps.retrieval.vector.main import VectorItem, SearchResult, GetResult
 from open_webui.config import (
@@ -18,128 +18,46 @@ from open_webui.config import (
     QDRANT_URL,
 )
 
+NO_LIMIT = 999999999
+
 
 class QdrantClient:
     def __init__(self):
+        self.collection_prefix = "open-webui"
         self.client = Client(
             url=QDRANT_URL, api_key=QDRANT_API_KEY, timeout=int(QDRANT_TIMEOUT_SECONDS)
         )
 
-    def _result_to_get_result(self, result) -> GetResult:
-        ids = []
-        documents = []
-        metadatas = []
+    def _result_to_get_result(self, points) -> GetResult:
+        ids, documents, metadatas = [], [], []
 
-        # Iterate over the tuple of records
-        for record in result[0]:
-            ids.append([record.id])
-            documents.append([record.payload["text"]])
-            metadatas.append([record.payload["metadata"]])
+        for point in points:
+            ids.append(point.id)
+            documents.append(point.payload["text"])
+            metadatas.append(point.payload["metadata"])
 
-        return GetResult(
-            **{
-                "ids": ids,
-                "documents": documents,
-                "metadatas": metadatas,
-            }
-        )
+        return GetResult(ids=[ids], documents=[documents], metadatas=[metadatas])
 
-    def _result_to_search_result(self, result) -> SearchResult:
-        ids = []
-        distances = []
-        documents = []
-        metadatas = []
+    def _result_to_search_result(self, points) -> SearchResult:
+        ids, distances, documents, metadatas = [], [], [], []
 
-        for point in result.points:
-            ids.append([point.id])
-            distances.append([point.score])
-            documents.append([point.payload["text"]])
-            metadatas.append([point.payload["metadata"]])
+        for point in points:
+            ids.append(point.id)
+            distances.append(point.score)
+            documents.append(point.payload["text"])
+            metadatas.append(point.payload["metadata"])
 
         return SearchResult(
-            **{
-                "ids": ids,
-                "distances": distances,
-                "documents": documents,
-                "metadatas": metadatas,
-            }
+            ids=[ids], distances=[distances], documents=[documents], metadatas=[metadatas]
         )
 
-    def has_collection(self, collection_name: str) -> bool:
-        # Check if the collection exists based on the collection name.
-        return self.client.collection_exists(collection_name=collection_name)
-
-    def delete_collection(self, collection_name: str):
-        # Delete the collection based on the collection name.
-        return self.client.delete_collection(collection_name=collection_name)
-
-    def search(
-        self, collection_name: str, vectors: list[list[float | int]], limit: int
-    ) -> Optional[SearchResult]:
-        # Search for the nearest neighbor items based on the vectors and return 'limit' number of results.
-        result = self.client.query_points(
-            collection_name=collection_name,
-            query=vectors,
-            limit=limit,
-            with_payload=True,
-        )
-
-        return self._result_to_search_result(result)
-
-    def query(
-        self, collection_name: str, filter: dict, limit: Optional[int] = None
-    ) -> Optional[GetResult]:
-        try:
-            if not self.client.collection_exists(collection_name=collection_name):
-                return None
-
-            # Build the conditions if a filter is provided.
-            qdrant_filter = None
-            if filter:
-                conditions = [
-                    FieldCondition(key=key, match=MatchValue(value=value))
-                    for key, value in filter.items()
-                ]
-                qdrant_filter = Filter(must=conditions)
-
-            points, _ = self.client.scroll(
-                collection_name=collection_name,
-                scroll_filter=qdrant_filter,
-                limit=limit or 1,
-            )
-
-            return self._result_to_get_result(points)
-
-        except Exception as e:
-            print(f"Error querying Qdrant: {e}")
-            return None
-
-    def get(self, collection_name: str) -> Optional[GetResult]:
-        points = self.client.count(
-            collection_name=collection_name,
-        )
-        if points.count:
-            # Get all the items in the collection.
-            result = self.client.scroll(
-                collection_name=collection_name,
-                with_payload=True,
-                limit=points.count,
-            )
-
-            return self._result_to_get_result(result)
-
-        return None
-
-    def insert(self, collection_name: str, items: list[VectorItem]):
-        return self.upsert(collection_name=collection_name, items=items)
-
-    def upsert(self, collection_name: str, items: list[VectorItem]):
-        # Update the items in the collection, if the items are not present, insert them. If the collection does not exist, it will be created.
-        if not self.client.collection_exists(collection_name=collection_name):
+    def _create_collection_if_not_exists(self, collection_name: str, dimension: int):
+        prefixed_name = f"{self.collection_prefix}_{collection_name}"
+        if not self.client.collection_exists(prefixed_name):
             self.client.create_collection(
-                collection_name=collection_name,
+                collection_name=prefixed_name,
                 vectors_config=VectorParams(
-                    size=len(items[0]["vector"]),
+                    size=dimension,
                     distance=Distance.COSINE,
                     multivector_config=models.MultiVectorConfig(
                         comparator=models.MultiVectorComparator.MAX_SIM
@@ -147,7 +65,8 @@ class QdrantClient:
                 ),
             )
 
-        points = [
+    def _create_points(self, items: list[VectorItem]):
+        return [
             PointStruct(
                 id=item["id"],
                 vector=item["vector"],
@@ -156,10 +75,66 @@ class QdrantClient:
             for item in items
         ]
 
-        return self.client.upsert(
-            collection_name=collection_name,
-            points=points,
+    def has_collection(self, collection_name: str) -> bool:
+        return self.client.collection_exists(f"{self.collection_prefix}_{collection_name}")
+
+    def delete_collection(self, collection_name: str):
+        return self.client.delete_collection(f"{self.collection_prefix}_{collection_name}")
+
+    def search(
+        self, collection_name: str, vectors: list[list[float | int]], limit: int
+    ) -> Optional[SearchResult]:
+        limit = limit or NO_LIMIT
+        prefixed_name = f"{self.collection_prefix}_{collection_name}"
+        query_response = self.client.query_points(
+            collection_name=prefixed_name,
+            query=vectors[0],
+            limit=limit,
+            with_payload=True,
         )
+        return self._result_to_search_result(query_response.points)
+
+    def query(
+        self, collection_name: str, filter: dict, limit: Optional[int] = None
+    ) -> Optional[GetResult]:
+        if not self.has_collection(collection_name):
+            return None
+        
+        try:
+            limit = limit or NO_LIMIT
+            field_conditions = [
+                FieldCondition(key=f"metadata.{key}", match=MatchValue(value=value))
+                for key, value in filter.items()
+            ]
+            
+            prefixed_name = f"{self.collection_prefix}_{collection_name}"
+            query_response = self.client.query_points(
+                collection_name=prefixed_name,
+                query_filter=Filter(must=field_conditions),
+                limit=limit,
+            )
+            return self._result_to_get_result(query_response.points)
+        except Exception as e:
+            print(f"Error querying Qdrant: {e}")
+            return None
+
+    def get(self, collection_name: str) -> Optional[GetResult]:
+        prefixed_name = f"{self.collection_prefix}_{collection_name}"
+        points = self.client.query_points(
+            collection_name=prefixed_name,
+            limit=NO_LIMIT,
+            with_payload=True,
+        )
+        return self._result_to_get_result(points.points)
+
+    def insert(self, collection_name: str, items: list[VectorItem]):
+        self.upsert(collection_name, items)
+
+    def upsert(self, collection_name: str, items: list[VectorItem]):
+        self._create_collection_if_not_exists(collection_name, len(items[0]["vector"]))
+        points = self._create_points(items)
+        prefixed_name = f"{self.collection_prefix}_{collection_name}"
+        return self.client.upsert(prefixed_name, points)
 
     def delete(
         self,
@@ -167,25 +142,26 @@ class QdrantClient:
         ids: Optional[list[str]] = None,
         filter: Optional[dict] = None,
     ):
-        # Delete the items from the collection based on the ids.
-        if ids:
-            selector = ids
-        elif filter:
-            conditions = [
-                FieldCondition(key=key, match=MatchValue(value=value))
-                for key, value in filter.items()
-            ]
-            selector = Filter(must=conditions)
+        field_conditions = [
+            FieldCondition(
+                key="metadata.id",
+                match=MatchValue(value=id_value),
+            )
+            for id_value in ids
+        ] if ids else [
+            FieldCondition(
+                key=f"metadata.{key}",
+                match=MatchValue(value=value),
+            )
+            for key, value in filter.items()
+        ]
 
         return self.client.delete(
-            collection_name=collection_name,
-            points_selector=selector,
+            collection_name=f"{self.collection_prefix}_{collection_name}",
+            points_selector=Filter(must=field_conditions),
         )
 
     def reset(self):
-        # Resets the database. This will delete all collections and item entries.
-
-        collection_response = self.client.get_collections()
-
-        for collection in collection_response.collections:
-            self.client.delete_collection(collection_name=collection.name)
+        for collection in self.client.get_collections().collections:
+            if collection.name.startswith(self.collection_prefix):
+                self.client.delete_collection(collection_name=collection.name)
